email fixed

+++ database/migrations/2020_08_28_151226_create_two_f_a_codes_table.php
@@ -15,11 +15,10 @@ class CreateTwoFACodesTable extends Migration
     {
         Schema::create('two_f_a_codes', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('code');
             $table->timestamps();
         });
-
         Schema::table('two_f_a_codes', function (Blueprint $table) {
             $table->foreign('user_id')->references('id')
                 ->on('users')->onDelete('cascade');
